исправление ошибок по отправке формы
